users can only view, update and delete their own notes

+++ controllers/posts/deletePost.js
@@ -5,10 +5,14 @@ const deletePost = async (req, res) => {
     const { _id } = req.params
     const post = await Post.findById({ _id })
     if (post) {
-      await Post.findByIdAndDelete({ _id })
-      res
-        .status(200)
-        .send({ message: "Post deleted successfully", deleted_post: post })
+      if (post.username === req.body.username) {
+        await Post.findByIdAndDelete({ _id })
+        res
+          .status(200)
+          .send({ message: "Post deleted successfully", deleted_post: post })
+      } else {
+        res.status(500).send({ error: "You are not authorized!" })
+      }
     } else {
       res.status(400).send({ error: "The post doesn't exist." })
     }
